Change all integer types to Long and all floating point types to Double

+++ src/main/java/com/stripe/model/EvidenceDetails.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @EqualsAndHashCode(callSuper = false)
 public final class EvidenceDetails extends StripeObject {
   protected Long dueBy;
-  protected Integer submissionCount;
+  protected Long submissionCount;
 }
